I made 2 changes
1. Hex Files Folder

I figured it would be a good idea to store the hex files for the game here, to reduce the need to copy and paste the python code into the micropython editor.

2. Flappy Sprite

Collision Detection Added
Score Counter Added
Game Over Added

The game is still a work in progress but it's definitely more robust than previous versions!